Changes to PreBuildingContext & integration with rendering

+++ common/src/main/java/me/cg360/mod/bridging/util/GameSupport.java
@@ -15,13 +15,13 @@ public class GameSupport {
 
     public static final double TRAPDOOR_HEIGHT = 3/16d;
     public static final double SLAB_HEIGHT = 8/16d;
-    public static final double MAXIMUM_PLACE_REACH = 4.5d;
+    public static final float MAXIMUM_PLACE_REACH = 4.5f;
 
-    public static double getReach() {
+    public static float getReach() {
         if(Minecraft.getInstance().player == null)
             return MAXIMUM_PLACE_REACH;
 
-        return Minecraft.getInstance().player.blockInteractionRange();
+        return (float) Minecraft.getInstance().player.blockInteractionRange();
     }
 
     public static boolean isControllerCrouching() {
